refactor: users앱 이름 auth로 수정
- domain도 users -> auth로 수정
- mysql내부의 장고관련 테이블 속성들도 수정
