+++ auth/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from users.views import SignUpView, SignInView
+from auth.views import SignUpView, SignInView
 
 urlpatterns = [
     path('/signup', SignUpView.as_view()),
