+++ wish_korea/settings.py
@@ -40,7 +40,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'corsheaders',
-    'users',
+    'auth',
     'products',
     'orders'
 ]
