+++ wish_korea/urls.py
@@ -16,7 +16,7 @@ Including another URLconf
 from django.urls import path, include
 
 urlpatterns = [
-    path('users', include('users.urls')),
+    path('auth', include('auth.urls')),
     path('products', include('products.urls')),
     path('orders', include('orders.urls')),
 ]
